+++ mentorship ios/Utility/URLStringConstants.swift
@@ -1,0 +1,11 @@
+//
+//  URLConstants.swift
+//  mentorship ios
+//
+//  Created by Yugantar Jain on 05/06/20.
+//  Copyright © 2020 Yugantar Jain. All rights reserved.
+//
+
+struct URLStringConstants {
+    static let login: String = "https://mentorship-backend-temp.herokuapp.com/login"
+}
